requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
Django==4.1.7
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
forex-python==1.8
idna==3.4
importlib-metadata==6.1.0
jsonpickle==3.0.1
Markdown==3.4.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
scraper==0.0.0
simplejson==3.18.4
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
validators==0.20.0
zipp==3.15.0
